Autoclose timers and refactor code
